refactor(File,Rank Info): file, rank enum util로 패키지 변경
