+++ src/main/java/view/GameCommand.java
@@ -2,8 +2,8 @@ package view;
 
 import java.util.Arrays;
 import java.util.List;
-import view.dto.FileInfo;
-import view.dto.RankInfo;
+import util.FileConverter;
+import util.RankConverter;
 
 public enum GameCommand {
     START("start"), END("end"), MOVE("move");
@@ -29,24 +29,24 @@ public enum GameCommand {
     public static int toSourceFileValue(List<String> inputCommand) {
         char file = inputCommand.get(SOURCE)
                 .charAt(FILE_INDEX);
-        return FileInfo.from(file).getValue();
+        return FileConverter.from(file).getValue();
     }
 
     public static int toSourceRankValue(List<String> inputCommand) {
         char rank = inputCommand.get(SOURCE)
                 .charAt(RANK_INDEX);
-        return RankInfo.from(rank).getValue();
+        return RankConverter.from(rank).getValue();
     }
 
     public static int toDestinationFileValue(List<String> inputCommand) {
         char file = inputCommand.get(DESTINATION)
                 .charAt(FILE_INDEX);
-        return FileInfo.from(file).getValue();
+        return FileConverter.from(file).getValue();
     }
 
     public static int toDestinationRankValue(List<String> inputCommand) {
         char rank = inputCommand.get(DESTINATION)
                 .charAt(RANK_INDEX);
-        return RankInfo.from(rank).getValue();
+        return RankConverter.from(rank).getValue();
     }
 }
